requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
lxml==5.1.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
flasgger==0.9.7.1
apispec==6.3.0
PyYAML==6.0.1
PyJWT==2.8.0
cryptography==41.0.7
pydantic==2.5.3
humanize==4.9.0
psutil==5.9.6
APScheduler==3.10.4

# E-Claim Import Dependencies
pandas==2.1.4
xlrd==2.0.1
openpyxl==3.1.2
psycopg2-binary==2.9.9
pymysql==1.1.0
sqlalchemy==2.0.23

# Machine Learning
scikit-learn==1.4.0
joblib==1.3.2
